requirements.txt
streamlit
langchain_community
openai
wikipedia
numexpr
google-search-results